requirements.txt
requests
Pillow
bgg-api